refactor: update file  input
